// infrastructure/kubernetes/generatedCode/crds/argoproj/v1alpha1/workflowTemplate.ts
// *** WARNING: this file was generated by crd2pulumi. ***
// *** Do not edit by hand unless you're certain you know what you are doing! ***

import * as pulumi from "@pulumi/pulumi";
import * as utilities from "../../utilities";

import { ObjectMeta } from "../../meta/v1";

export class WorkflowTemplate extends pulumi.CustomResource {
	/**
	 * Get an existing WorkflowTemplate resource's state with the given name, ID, and optional extra
	 * properties used to qualify the lookup.
	 *
	 * @param name The _unique_ name of the resulting resource.
	 * @param id The _unique_ provider ID of the resource to lookup.
	 * @param opts Optional settings to control the behavior of the CustomResource.
	 */
	public static get(
		name: string,
		id: pulumi.Input<pulumi.ID>,
		opts?: pulumi.CustomResourceOptions,
	): WorkflowTemplate {
		return new WorkflowTemplate(name, undefined as any, { ...opts, id: id });
	}

	/** @internal */
	public static readonly __pulumiType =
		"kubernetes:argoproj.io/v1alpha1:WorkflowTemplate";

	/**
	 * Returns true if the given object is an instance of WorkflowTemplate.  This is designed to work even
	 * when multiple copies of the Pulumi SDK have been loaded into the same process.
	 */
	public static isInstance(obj: any): obj is WorkflowTemplate {
		if (obj === undefined || obj === null) {
			return false;
		}
		return obj["__pulumiType"] === WorkflowTemplate.__pulumiType;
	}

	public readonly apiVersion!: pulumi.Output<
		"argoproj.io/v1alpha1" | undefined
	>;
	public readonly kind!: pulumi.Output<"WorkflowTemplate" | undefined>;
	public readonly metadata!: pulumi.Output<ObjectMeta>;
	public readonly spec!: pulumi.Output<{ [key: string]: any }>;

	/**
	 * Create a WorkflowTemplate resource with the given unique name, arguments, and options.
	 *
	 * @param name The _unique_ name of the resource.
	 * @param args The arguments to use to populate this resource's properties.
	 * @param opts A bag of options that control this resource's behavior.
	 */
	constructor(
		name: string,
		args?: WorkflowTemplateArgs,
		opts?: pulumi.CustomResourceOptions,
	) {
		let resourceInputs: pulumi.Inputs = {};
		opts = opts || {};
		if (!opts.id) {
			resourceInputs["apiVersion"] = "argoproj.io/v1alpha1";
			resourceInputs["kind"] = "WorkflowTemplate";
			resourceInputs["metadata"] = args ? args.metadata : undefined;
			resourceInputs["spec"] = args ? args.spec : undefined;
		} else {
			resourceInputs["apiVersion"] = undefined /*out*/;
			resourceInputs["kind"] = undefined /*out*/;
			resourceInputs["metadata"] = undefined /*out*/;
			resourceInputs["spec"] = undefined /*out*/;
		}
		opts = pulumi.mergeOptions(utilities.resourceOptsDefaults(), opts);
		super(WorkflowTemplate.__pulumiType, name, resourceInputs, opts);
	}
}

/**
 * The set of arguments for constructing a WorkflowTemplate resource.
 */
export interface WorkflowTemplateArgs {
	apiVersion?: pulumi.Input<"argoproj.io/v1alpha1">;
	kind?: pulumi.Input<"WorkflowTemplate">;
	metadata?: pulumi.Input<ObjectMeta>;
	spec?: pulumi.Input<{ [key: string]: any }>;
}